requirements.txt
matplotlib==3.3.4
numpy==1.22.0
opencv_python==4.5.5.62
Pillow==9.0.0
pip==21.3.1
pytesseract==0.3.8
pytube==11.0.2
setuptools==58.2.0
string_color==1.2.1
thefuzz==0.19.0
wheel==0.34.2
